// apps/commerce/src/enums/order-status.enum.ts
export enum OrderStatus {
    PENDING_KIOSK_CONFIRMATION = 'PENDING_KIOSK_CONFIRMATION',
    ACCEPTED = 'ACCEPTED',
    REJECTED = 'REJECTED',
    PENDING_LOGISTICS_PRICE = 'PENDING_LOGISTICS_PRICE',
    READY_FOR_PAYMENT = 'READY_FOR_PAYMENT',
    PAID = 'PAID',
    CANCELLED = 'CANCELLED',
    AUTO_REJECTED_TIMEOUT = 'AUTO_REJECTED_TIMEOUT',
    CANCEL_REQUESTED = 'CANCEL_REQUESTED',
    REFUND_PENDING = 'REFUND_PENDING',
    REFUNDED = 'REFUNDED'
}